ref(utils): hoistable functions
This commit refactors the helpers in util.js to use the classic
hoistable JavaScript function pattern instead of const+arrow. This
is needed because our build process just bundles files together in
a random order. (So a function could be defined in arbitrarty order
in the final server.js file)