fix(enemy.js): fix enemy collision teleportation
Moves the y-direction collision checks before the x-direction checks in updateSpeed() and flips a dirty flag to true if there is a top collision. Then the x-direction checks will only move the enemy to the side of the obstacle if there hasn't already been a top collision. Prioritizes y-collisions but that shouldn't be a problem.