feat(enemy): enemy death animations
Adds stylized exploding fade out animation to enemies on death. Adjust deathAnimationTotalFrames in sprites.js to change the length of the animation. Adds hex to RGB converter to client.js so we can adjust the alpha (transparency) of the animation.

TODO: make an Entity class on the client so we can use static functions like hexToRGB from there (low priority)